company_archive_mvp_chroma/requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
openai==1.37.1
pymupdf==1.24.10
pandas==2.2.2
tqdm==4.66.4
sqlite-utils==3.36
chromadb==0.5.5
httpx==0.27.2